+++ lib/anycable/rails/compatibility.rb
@@ -34,6 +34,7 @@ module AnyCable
         diff = instance_variables - was_ivars
 
         if self.class.respond_to?(:channel_state_attributes)
+          diff.delete(:@__istate__)
           diff.delete_if { |ivar| self.class.channel_state_attributes.include?(:"#{ivar.to_s.sub(/^@/, "")}") }
         end